requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.9.0
sqlalchemy[asyncio]==2.0.0
alembic==1.13.0
asyncpg==0.29.0
redis[hiredis]==5.1.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
GitPython==3.1.40
PyYAML==6.0.2
loguru==0.7.2
aiohttp==3.9.5
slack-bolt==1.18.0
slack-sdk==3.26.0
python-telegram-bot==20.7
twilio==8.11.0
aiosmtplib==3.0.1
signal-cli-rest-api==0.9.0
aioredis==2.0.1